+++ requests.go
@@ -14,7 +14,6 @@ package requests
 
 import (
 	"bytes"
-	"compress/gzip"
 	"crypto/tls"
 	"encoding/json"
 	"fmt"
@@ -25,26 +24,17 @@ import (
 	"net/http/cookiejar"
 	"net/http/httputil"
 	"net/url"
-	"os"
 	"strings"
 	"time"
 )
 
-var VERSION string = "0.8"
-
 type Request struct {
-	httpreq *http.Request
-	Header  *http.Header
-	Client  *http.Client
-	debug   bool
-	Cookies []*http.Cookie
-}
-
-type Response struct {
-	R       *http.Response
-	content []byte
-	text    string
-	req     *Request
+	httpreq     *http.Request
+	Header      *http.Header
+	Client      *http.Client
+	debug       bool
+	respHandler func(*Response)
+	Cookies     []*http.Cookie
 }
 
 type Header map[string]string
@@ -62,16 +52,7 @@ type Auth []string
 func Requests(method string) *Request {
 
 	req := new(Request)
-
-	req.httpreq = &http.Request{
-		Method:     strings.ToUpper(method),
-		Header:     make(http.Header),
-		Proto:      "HTTP/1.1",
-		ProtoMajor: 1,
-		ProtoMinor: 1,
-	}
-	req.Header = &req.httpreq.Header
-	req.httpreq.Header.Set("User-Agent", "Go-Requests "+VERSION)
+	req.reset()
 
 	req.Client = &http.Client{}
 
@@ -84,54 +65,22 @@ func Requests(method string) *Request {
 	return req
 }
 
-// Get ,req.Get
-
-func Get(origurl string, args ...interface{}) (resp *Response, err error) {
-	// call request Get
-	resp, err = Requests("GET").Run(origurl, args...)
-	return resp, err
-}
-
-// handle URL params
-func buildURLParams(userURL string, params ...map[string]string) (string, error) {
-	parsedURL, err := url.Parse(userURL)
-
-	if err != nil {
-		return "", err
+func (req *Request) reset() {
+	req.httpreq = &http.Request{
+		Method:     "GET",
+		Header:     make(http.Header),
+		Proto:      "HTTP/1.1",
+		ProtoMajor: 1,
+		ProtoMinor: 1,
 	}
-
-	parsedQuery, err := url.ParseQuery(parsedURL.RawQuery)
-
-	if err != nil {
-		return "", nil
-	}
-
-	for _, param := range params {
-		for key, value := range param {
-			parsedQuery.Add(key, value)
-		}
-	}
-	return addQueryParams(parsedURL, parsedQuery), nil
-}
-
-func addQueryParams(parsedURL *url.URL, parsedQuery url.Values) string {
-	if len(parsedQuery) > 0 {
-		return strings.Join([]string{strings.Replace(parsedURL.String(), "?"+parsedURL.RawQuery, "", -1), parsedQuery.Encode()}, "?")
-	}
-	return strings.Replace(parsedURL.String(), "?"+parsedURL.RawQuery, "", -1)
-}
-
-func (req *Request) SetDebug(debug bool) *Request {
-	req.debug = debug
-	return req
+	req.Header = &req.httpreq.Header
+	req.httpreq.Header.Set("User-Agent", "Go-Requests "+VERSION)
 }
 
 func (req *Request) RequestDebug() {
-
 	if req.debug {
 		return
 	}
-
 	fmt.Println("===========Go RequestDebug ============")
 
 	message, err := httputil.DumpRequestOut(req.httpreq, false)
@@ -194,124 +143,24 @@ func (req *Request) Proxy(proxyurl string) {
 	}
 }
 
-/**************/
-func (resp *Response) ResponseDebug() {
-	if resp.req.debug {
-		return
-	}
-
-	fmt.Println("===========Go ResponseDebug ============")
-
-	message, err := httputil.DumpResponse(resp.R, false)
-	if err != nil {
-		return
-	}
-
-	fmt.Println(string(message))
-
-}
-
-func (resp *Response) Content() []byte {
-	var err error
-	if resp.content != nil {
-		return resp.content
-	}
-	defer resp.R.Body.Close()
-
-	var Body = resp.R.Body
-	if resp.R.Header.Get("Content-Encoding") == "gzip" && resp.req.Header.Get("Accept-Encoding") != "" {
-		// fmt.Println("gzip")
-		reader, err := gzip.NewReader(Body)
-		if err != nil {
-			return nil
-		}
-		Body = reader
-	}
-
-	resp.content, err = ioutil.ReadAll(Body)
-	if err != nil {
-		return nil
-	}
-
-	return resp.content
-}
-
-func (resp *Response) Text() string {
-	if resp.content == nil {
-		resp.Content()
-	}
-	resp.text = string(resp.content)
-	return resp.text
-}
-
-func (resp *Response) SaveFile(filename string) error {
-	if resp.content == nil {
-		resp.Content()
-	}
-	f, err := os.Create(filename)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	_, err = f.Write(resp.content)
-	f.Sync()
-
-	return err
-}
-
-func (resp *Response) Json(v interface{}) error {
-	if resp.content == nil {
-		resp.Content()
-	}
-	return json.Unmarshal(resp.content, v)
-}
-
-func (resp *Response) Cookies() (cookies []*http.Cookie) {
-	httpreq := resp.req.httpreq
-	client := resp.req.Client
-
-	cookies = client.Jar.Cookies(httpreq.URL)
-
-	return cookies
-
-}
-
-/**************post*************************/
-// call req.Post ,only for easy
-func Post(origurl string, args ...interface{}) (resp *Response, err error) {
-	resp, err = Requests("POST").Run(origurl, args...)
-	return
-}
-
-// Put
-func Put(origurl string, args ...interface{}) (resp *Response, err error) {
-	resp, err = Requests("PUT").Run(origurl, args...)
-	return
-}
-
-// Delete
-func Delete(origurl string, args ...interface{}) (resp *Response, err error) {
-	resp, err = Requests("DELETE").Run(origurl, args...)
-	return
-}
-
-// Patch
-func Patch(origurl string, args ...interface{}) (resp *Response, err error) {
-	resp, err = Requests("PATCH").Run(origurl, args...)
-	return
-}
-
 // cleanup
 func (req *Request) cleanup() {
 	req.httpreq.Body = nil
 	req.httpreq.GetBody = nil
 	req.httpreq.ContentLength = 0
-	req.Header = &http.Header{}
+	// req.Header = &http.Header{}
+	// req.Header = &req.httpreq.Header
+
 	//reset Cookies,
 	//Client.Do can copy cookie from client.Jar to req.Header
 	delete(req.httpreq.Header, "Cookie")
 	// req.ClearCookies()
+}
+
+// SetRespHandler
+func (req *Request) SetRespHandler(fn func(*Response)) *Request {
+	req.respHandler = fn
+	return req
 }
 
 // SetMethod
@@ -329,7 +178,7 @@ func (req *Request) SetHeader(key, value string) *Request {
 // Post -
 func (req *Request) Run(origurl string, args ...interface{}) (resp *Response, err error) {
 	// cleanup
-	req.cleanup()
+	// req.cleanup()
 
 	// set params ?a=b&b=c
 	//set Header
@@ -404,6 +253,10 @@ func (req *Request) Run(origurl string, args ...interface{}) (resp *Response, er
 	resp.req = req
 	resp.ResponseDebug()
 	resp.Content()
+	req.reset()
+	if req.respHandler != nil {
+		req.respHandler(resp)
+	}
 	return resp, nil
 }
 
@@ -418,9 +271,6 @@ func (req *Request) setBodyForms(Forms url.Values) {
 func (req *Request) setBodyBytes(data []byte) {
 	req.httpreq.Body = ioutil.NopCloser(bytes.NewReader(data))
 	req.httpreq.ContentLength = int64(len(data))
-}
-func (req *Request) setBodyRawBytes(read io.ReadCloser) {
-	req.httpreq.Body = read
 }
 
 // upload file and form
@@ -477,14 +327,4 @@ func (req *Request) buildJSON(data interface{}) []byte {
 
 	// fmt.Printf("a1=%#v,jsons=%#v\nahui\n", data, string(jsonBytes))
 	return jsonBytes
-}
-
-// open file for post upload files
-
-func openFile(filename string) *os.File {
-	r, err := os.Open(filename)
-	if err != nil {
-		panic(err)
-	}
-	return r
 }
