+++ response.go
@@ -1,0 +1,109 @@
+/* Copyright（2） 2018 by  asmcos and ahuigo .
+Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+     http://www.apache.org/licenses/LICENSE-2.0
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
+
+package requests
+
+import (
+	"compress/gzip"
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"net/http/httputil"
+	"os"
+)
+
+type Response struct {
+	R       *http.Response
+	content []byte
+	text    string
+	req     *Request
+}
+
+func (resp *Response) ResponseDebug() {
+	if resp.req.debug {
+		return
+	}
+	fmt.Println("===========Go ResponseDebug ============")
+
+	message, err := httputil.DumpResponse(resp.R, false)
+	if err != nil {
+		return
+	}
+
+	fmt.Println(string(message))
+}
+
+func (resp *Response) Content() []byte {
+	var err error
+	if resp.content != nil {
+		return resp.content
+	}
+	defer resp.R.Body.Close()
+
+	var Body = resp.R.Body
+	if resp.R.Header.Get("Content-Encoding") == "gzip" && resp.req.Header.Get("Accept-Encoding") != "" {
+		reader, err := gzip.NewReader(Body)
+		if err != nil {
+			return nil
+		}
+		Body = reader
+	}
+
+	resp.content, err = ioutil.ReadAll(Body)
+	if err != nil {
+		return nil
+	}
+
+	return resp.content
+}
+
+func (resp *Response) Text() string {
+	if resp.content == nil {
+		resp.Content()
+	}
+	resp.text = string(resp.content)
+	return resp.text
+}
+
+func (resp *Response) SaveFile(filename string) error {
+	if resp.content == nil {
+		resp.Content()
+	}
+	f, err := os.Create(filename)
+	if err != nil {
+		return err
+	}
+	defer f.Close()
+
+	_, err = f.Write(resp.content)
+	f.Sync()
+
+	return err
+}
+
+func (resp *Response) Json(v interface{}) error {
+	if resp.content == nil {
+		resp.Content()
+	}
+	return json.Unmarshal(resp.content, v)
+}
+
+func (resp *Response) Cookies() (cookies []*http.Cookie) {
+	httpreq := resp.req.httpreq
+	client := resp.req.Client
+
+	cookies = client.Jar.Cookies(httpreq.URL)
+
+	return cookies
+
+}
